+++ ruby/red-gandiva/lib/gandiva/expression-builder/record.rb
@@ -16,40 +16,22 @@
 # under the License.
 
 module Gandiva
-  class IfNodeQuery
-    def initialize(condition)
-      @condition_builders = [condition]
-      @then_builders = []
-    end
-
-    def then(builder)
-      @then_builders << builder
-      self
-    end
-
-    def elsif(builder)
-      @condition_builders << builder
-      self
-    end
-
-    def else(builder)
-      to_expression_builder(builder)
-    end
-
-    private
-
-    def to_expression_builder(builder)
-      node_size = @condition_builders.size - 1
-      (0..node_size).reverse_each do |i|
-        builder = build_if_expression_builder(builder, i)
+  class ExpressionBuilder
+    class Record
+      def initialize(schema)
+        @schema = schema
       end
-      builder
-    end
 
-    def build_if_expression_builder(builder, i)
-      Gandiva::IfExpressionBuilder.new(@condition_builders[i],
-                                       @then_builders[i],
-                                       builder)
+      def respond_to_missing?(name, include_private)
+        return true if @schema[name]
+        super
+      end
+
+      def method_missing(name, *args)
+        field = @schema[name]
+        return Field.new(field) if field
+        super
+      end
     end
   end
 end
